+++ packages/react-app-super-template/src/keycloak.ts
@@ -1,3 +1,3 @@
 import Keycloak from 'keycloak-js';
 
-export default Keycloak(globalThis.ENV?.KEYCLOAK);
+export default new Keycloak(globalThis.ENV?.KEYCLOAK);